+++ CHANGELOG.md
@@ -8,9 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
-- **Configfix integration** for configuration repair analysis
-
 ### Changed
+
+- Revised export of Docker images, allowing to easily create GitHub releases with all images
 
 ### Removed